+++ R/realtime_plot.R
@@ -85,8 +85,8 @@ plot.realtime <- function(x = NULL, Parameter = c("Flow", "Level"), ...) {
     mgp = c(3.1, 0.4, 0),
     las = 1,
     tck = -.01,
-    xaxs = "i",
-    yaxs = "i"
+    xaxs = "r",
+    yaxs = "r"
   )
 
   graphics::plot(